refactor: OCR 에서 응답 파싱 책임 분리
